\|/ |-| /-\ ‾|‾ ' _|‾   ‾|‾ |-| | _|‾ ?
-Radiohead

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teleOps/MainTele.java
@@ -107,7 +107,7 @@ public class MainTele extends OpMode {
         else if (gamepad2.right_bumper) board.theHookBringsYouBack(-1);
         else board.theHookBringsYouBack(0);
 
-        intakeLiftPos += (gamepad1.left_trigger - gamepad1.right_trigger) * 100;
+        intakeLiftPos += (int) (gamepad1.left_trigger - gamepad1.right_trigger) * 100;
         intakeLiftPos = max(intakeLiftPos, UpAndDownServoLift.UP.getPos());
         intakeLiftPos = min(intakeLiftPos, UpAndDownServoLift.DOWN.getPos());
